+++ public/css/app.css
@@ -1650,3 +1650,31 @@ h1.checkout-heading {
   font-weight: normal;
 }
 
+.register-info-section .store-section {
+  display: grid;
+  margin: 80px auto 80px;
+}
+
+.register-info-section .delete-shop-button {
+  border-color: #a91212;
+  padding: 10px 32px !important;
+  background-color: #b73131;
+  color: #e9e9e9;
+  border-radius: 5px;
+  padding: 12px 50px;
+}
+
+.register-info-section .delete-shop-button:hover {
+  background-color: #c54141 !important;
+}
+
+.register-info-section .delete-shop-button:disabled {
+  cursor: not-allowed;
+  background: whitesmoke;
+  color: rgba(0, 0, 0, 0.25);
+  background-color: #f5f5f5;
+  border-color: #d9d9d9;
+  text-shadow: none;
+  box-shadow: none;
+}
+
