+++ resources/sass/pages/register-store.scss
@@ -94,4 +94,32 @@
     .label {
       font-weight: normal;
     }
+
+    .store-section {
+      display: grid;
+    
+      margin: 80px auto 80px;
+	}
+	
+	.delete-shop-button {
+		border-color: #a91212;
+		padding: 10px 32px !important;
+		background-color: #b73131;
+        color: $white;
+        border-radius: 5px;
+        padding: 12px 50px;
+        &:hover {
+            background-color: #c54141 !important;
+        }
+        &:disabled {
+          cursor: not-allowed;
+          background: whitesmoke;
+          color: rgba(0,0,0,.25);
+          background-color: #f5f5f5;
+          border-color: #d9d9d9;
+          text-shadow: none;
+		  box-shadow: none;
+        }
+
+    }
 }
